Fix: adjust Carousel timing, import trimmed clip
Adjust AliceCarousel timing to better match video lengths while
auto-playing. Trim second video of carousel, as it was much longer than
the rest, skewing the carousel's timing. Place state-change for Sliding-
Menu of NavBar for when NavBar disappears when scrolling up; as NavBar
disappears, so does a potentially open Sliding-Menu.