Feat: import image, animate backside of cover
Import image for backside of book cover, so when flipped it gives the
illusion of a book being opened. Backside of cover made by cropping
center of front cover and importing as new, separate image. Create
element for cover-backside, use same state as cover flip. Use same
sliding animation as cover upon load. When cover is flipped, z-index of
backside is changed as soon as the cover is turned to a 90deg angle from
user's perspective, creating illusion of open book.